Fix ubsan warning
src/gui-qt/metatiles/mttileset/mttilesetgraphicsscenes.cpp:24:17: runtime error: implicit conversion from type 'int' of value 256 (32-bit, signed) to type 'uint8_t' (aka 'unsigned char') changed the value to 0 (8-bit, unsigned)

+++ src/gui-qt/metatiles/mttileset/mttilesetgraphicsscenes.cpp
@@ -19,7 +19,7 @@ const grid<uint8_t> MtTilesetGraphicsScene::_grid = []() {
 
     UnTech::grid<uint8_t> grid(MT::TILESET_WIDTH, MT::TILESET_HEIGHT);
 
-    uint8_t i = 0;
+    unsigned i = 0;
     for (uint8_t& cell : grid) {
         cell = i++;
     }
